AImusicbot/requirements.txt
discord.py==2.3.2
yt-dlp==2025.07.02.233055.dev0
PyNaCl==1.5.0
python-dotenv
google-api-python-client
transformers
torch
numba
numpy
tflite_runtime
pycoral
transformers
huggingface-hub
requests
llama-cpp-python[cuda] # Or llama-cpp-python if no CUDA